separated lobbys
