+++ app/src/main/java/com/p2p/presentation/impostor/ImpostorActivity.kt
@@ -6,7 +6,8 @@ import androidx.activity.viewModels
 import com.p2p.presentation.basegame.GameActivity
 import com.p2p.presentation.impostor.create.CreateImpostorFragment
 import com.p2p.presentation.tuttifrutti.countdown.TuttiFruttiCountdownFragment
-import com.p2p.presentation.clientlobby.ClientLobbyFragment
+import com.p2p.presentation.lobby.ClientLobbyFragment
+import com.p2p.presentation.tuttifrutti.lobby.ImpostorClientLobbyFragment
 
 class ImpostorActivity : GameActivity<ImpostorSpecificGameEvent, ImpostorViewModel>() {
 
@@ -24,7 +25,7 @@ class ImpostorActivity : GameActivity<ImpostorSpecificGameEvent, ImpostorViewMod
 
     //TODO bren put impostor lobby
     override fun goToClientLobby() =
-        addFragment(ClientLobbyFragment.newInstance(viewModel), shouldAddToBackStack = false)
+        addFragment(ImpostorClientLobbyFragment.newInstance(), shouldAddToBackStack = false)
 
     companion object {
