+++ app/src/main/java/com/p2p/presentation/lobby/ClientLobbyViewModel.kt
@@ -1,4 +1,4 @@
-package com.p2p.presentation.clientlobby
+package com.p2p.presentation.lobby
 
 import com.p2p.presentation.base.BaseViewModel
