+++ app/src/main/java/com/p2p/presentation/lobby/lobbyEvent.kt
@@ -1,4 +1,4 @@
-package com.p2p.presentation.clientlobby
+package com.p2p.presentation.lobby
 
 sealed class LobbyEvent
